requirements.txt
psycopg2
django-environ
python-decouple
pytest